requirement.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
opencv-python>=4.5.0
Pillow>=8.0.0
clip-by-openai @ git+https://github.com/openai/CLIP.git
